+++ src/pike_memory.c
@@ -52,13 +52,13 @@ INLINE p_wchar2 *MEMCHR2(p_wchar2 *p, p_wchar2 c, ptrdiff_t e)
   return (p_wchar2 *)0;
 }
 
-void swap(char *a, char *b, INT32 size)
+void swap(char *a, char *b, size_t size)
 {
-  int tmp;
+  size_t tmp;
   char tmpbuf[1024];
   while(size)
   {
-    tmp=MINIMUM((long)sizeof(tmpbuf), size);
+    tmp = MINIMUM((size_t)sizeof(tmpbuf), size);
     MEMCPY(tmpbuf,a,tmp);
     MEMCPY(a,b,tmp);
     MEMCPY(b,tmpbuf,tmp);
@@ -68,7 +68,7 @@ void swap(char *a, char *b, INT32 size)
   }
 }
 
-void reverse(char *memory, INT32 nitems, INT32 size)
+void reverse(char *memory, size_t nitems, size_t size)
 {
 
 #define DOSIZE(X,Y)						\
